requirements/development.txt
# These packages will greatly aid development efforts, but
# do not need to be installed in a production environment.
coverage==3.6
django-coverage==1.2.2
django-debug-toolbar==0.9.4
django-extensions==1.1.1
pep8==1.4.5
pyflakes==0.7.2
selenium==2.32.0
Sphinx==1.2b1
Werkzeug==0.8.3